GoogleMaps Intent
Basically I created an Google Map Intent with all the functionality, since we are moving into using Google Accounts for everything, this will save all the information in your Google Account of any places you want to visit. I removed everything else that is not needed and updated the GoogleMaps API since it is required to add a billing account.